Added to the fitness function
The fitness now calculates the fitness based on how often objects are in
range of firing as well.
This calculation is pretty poor though, so we will probably need to
change it at some point.
Also added the ability for the user to affect the fitness, but it
depends on the UI to be implemented first.